requirements.txt
absl-py==0.9.0
astor==0.8.1
backcall==0.1.0
boto3==1.12.14
botocore==1.15.14
cachetools==4.1.0
certifi==2018.8.24
chardet==3.0.4
clang==11.0
Click==7.0
configparser==5.0.0
cvxpy==1.1.3
cycler==0.10.0
dataclasses==0.7
decorator==4.4.2
docker-pycreds==0.4.0
docutils==0.15.2
dominate==2.3.4
ecos==2.0.7.post1
filelock==3.0.12
flake8==3.8.3
future==0.18.2
gast==0.2.2
gitdb==4.0.4
GitPython==3.1.1
google-auth==1.17.2
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
gql==0.2.0
graphql-core==1.1
grpcio==1.29.0
h5py==2.10.0
idna==2.9
importlib-metadata==1.6.1
ipdb==0.13.2
ipython==7.13.0
ipython-genutils==0.2.0
jedi==0.16.0
jmespath==0.9.5
joblib==0.14.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
Markdown==3.2.2
matplotlib==3.2.1
mccabe==0.6.1
mpmath==1.1.0
numpy==1.18.1
nvidia-ml-py3==7.352.0
oauthlib==3.1.0
opt-einsum==3.2.1
osqp==0.6.1
pandas==1.0.3
parso==0.6.2
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.1
pipenv==2018.7.1
promise==2.3
prompt-toolkit==3.0.3
protobuf==3.12.2
psutil==5.7.0
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pyflakes==2.2.0
Pygments==2.5.2
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3.1
regex==2020.2.20
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.6
s3transfer==0.3.3
sacremoses==0.0.38
scipy==1.5.2
scs==2.1.2
sentencepiece==0.1.85
sentry-sdk==0.14.3
shortuuid==1.0.1
six==1.14.0
smmap==3.0.2
strconv==0.4.2
subprocess32==3.5.4
sympy==1.5.1
tensorboard==2.0.2
tensorflow-estimator==2.0.1
termcolor==1.1.0
tokenizers==0.5.2
torch==1.5.1
torchvision==0.6.1
tqdm==4.43.0
traitlets==4.3.3
urllib3==1.25.8
virtualenv==16.0.0
virtualenv-clone==0.3.0
wandb==0.8.32
watchdog==0.10.2
wcwidth==0.1.8
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.1.0
